external_api/pyproject.toml
[project]
name = "external-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastmcp>=2.14.2",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]
